Agentic-AI(Sun2-5)/Class_Work/Class_5/pyproject.toml
[project]
name = "class-5"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.3",
]
